:truck: Move version parsing into separate module
